Move configuration classes to a separate folder.
