+++ EntityFramework.PostgreSql/EntityFramework.PostgreSql.csproj
@@ -69,14 +69,14 @@
   <ItemGroup>
     <Compile Include="CustomHistoryContext.cs" />
     <Compile Include="PostgreSqlDbMigrationsConfiguration.cs" />
-    <Compile Include="IPostgreSqlConfiguration.cs" />
-    <Compile Include="IPostgreSqlSchemas.cs" />
+    <Compile Include="Configuration\IPostgreSqlConfiguration.cs" />
+    <Compile Include="Configuration\IPostgreSqlSchemas.cs" />
     <Compile Include="LowerCasePropertyNameConvention.cs" />
     <Compile Include="PostgreSqlDbContext.cs" />
     <Compile Include="PostgreSqlDbMigration.cs" />
-    <Compile Include="PostgreSqlConfiguration.cs" />
-    <Compile Include="PostgreSqlSchemaConfiguration.cs" />
-    <Compile Include="PostgreSqlSchemaConfigurationElement.cs" />
+    <Compile Include="Configuration\PostgreSqlConfiguration.cs" />
+    <Compile Include="Configuration\PostgreSqlSchemaConfiguration.cs" />
+    <Compile Include="Configuration\PostgreSqlSchemaConfigurationElement.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
