+++ EntityFramework.PostgreSql/app.config.transform
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <configuration>
   <configSections>
-    <section name="postgreSql" type="Affecto.EntityFramework.PostgreSql.PostgreSqlConfiguration, Affecto.EntityFramework.PostgreSql" requirePermission="false"/>
+    <section name="postgreSql" type="Affecto.EntityFramework.PostgreSql.Configuration.PostgreSqlConfiguration, Affecto.EntityFramework.PostgreSql" requirePermission="false"/>
   </configSections>
 
   <postgreSql>
